fix routing bug
