+++ src/App.css
@@ -1,0 +1,43 @@
+html {
+  box-sizing: border-box;
+  background-color: #3c99dc;
+  color: #fff;
+}
+
+*{
+  list-style-type: none;
+  text-decoration: none;
+}
+header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+h2 {
+  margin: 0 10%;
+}
+
+.intro img {
+  width: 50%;
+}
+
+img{
+  width:25%;
+}
+
+.intro {
+  display: flex;
+  grid-column: span 2;
+}
+
+.grid-container {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-template-rows: repeat(5, 1fr);
+  gap: 20px;
+}
+
+
+
+
+
