--- src/Components/Intro.js
@@ -1,9 +1,0 @@
-const Intro = () => (
-  <div>
-  <img src=""  alt=""></img>
-  <h3>The World</h3>
-  <span>195 countries</span>  
-  </div>
-)
-
-export default Intro
