+++ src/Components/homePage.js
@@ -1,69 +1,34 @@
-import { v4 as uuidv4 } from 'uuid';
+
 import {Link} from 'react-router-dom'
 import { useDispatch } from 'react-redux';
 import { FetchCountires } from '../Redux/DetailsPage/DetailsPage';
+import continentsData from '../Data/Data';
 
 
-const continentsData = [
-  {
-    id: uuidv4(),
-    name: 'Africa',
-    map: '',
-    image: '',
-  },
-  {
-    id: uuidv4(),
-    name: 'Asia',
-    map: '',
-    image: '/images/Asia.svg',
-  },
-  {
-    id: uuidv4(),
-    name: 'Europe',
-    map: '',
-    image: '/images/Europe.svg',
-  },
-  {
-    id: uuidv4(),
-    name: 'Oceania',
-    map: '',
-    image: '/images/Oceania.svg',
-  },
-  {
-    id: uuidv4(),
-    name: 'North America',
-    map: '',
-    image: '/images/North America.svg',
-  },
-  {
-    id: uuidv4(),
-    name: 'South America',
-    map: '',
-    image: '/images/South America.svg',
-  },
-  {
-    id: uuidv4(),
-    name: 'Antarctica',
-    map: '',
-    image: '/images/Antarctica.svg',
-  },
-];
+
 
 const Home = () => {
  const  dispatch = useDispatch()
 
 return(
-  <div>
-    {continentsData.map((continent) => (
-    <Link onClick={() => dispatch(FetchCountires(continent.name))} key = {continent.id} to={`/details/${continent.name}`}>
-        <ul>
-          <li>{continent.name}</li>
-          <li>{continent.image}</li>
-        </ul>
-      </Link>
-  
-  ))}
+  <div className='grid-container'>
+    <div className="intro">
+    <img src="./images/world.svg" alt=""></img>
+    <div>
+      <h3>The World</h3>
+      <span>195 countries</span>
+    </div>
   </div>
+      {continentsData.map((continent) => (
+        <Link onClick={() => dispatch(FetchCountires(continent.name))} key={continent.id} to={`/details/${continent.name}`}>
+          <ul>
+            <li>{continent.name}</li>
+            <li><img src={continent.image} alt="contImgae" /></li>
+          </ul>
+        </Link>
+
+      ))}
+    </div>
   
 )
 }
