+++ src/Data/Data.js
@@ -1,0 +1,48 @@
+import { v4 as uuidv4 } from 'uuid';
+
+const continentsData = [
+  {
+    id: uuidv4(),
+    name: 'africa',
+    map: '',
+    image: '/images/africa.svg',
+  },
+  {
+    id: uuidv4(),
+    name: 'Asia',
+    map: '',
+    image: './images/asia.svg',
+  },
+  {
+    id: uuidv4(),
+    name: 'Europe',
+    map: '',
+    image: '/images/europe.svg',
+  },
+  {
+    id: uuidv4(),
+    name: 'Oceania',
+    map: '',
+    image: '/images/oceania.svg',
+  },
+  {
+    id: uuidv4(),
+    name: 'North America',
+    map: '',
+    image: '/images/north-america.svg',
+  },
+  {
+    id: uuidv4(),
+    name: 'South America',
+    map: '',
+    image: '/images/south-america.svg',
+  },
+  {
+    id: uuidv4(),
+    name: 'Antarctica',
+    map: '',
+    image: '/images/Antarctica.svg',
+  },
+];
+
+export default continentsData
